Fixed a lot of clang compiler warnings.

+++ hurl_parse.c
@@ -47,7 +47,7 @@ int hurl_parse_url(char *url_ptr, HURLParsedURL **result) {
 		return 0;
 	}
 
-	(*result)->protocol = hurl_allocstrcpy(url, eof_protocol - url, 1);
+	(*result)->protocol = hurl_allocstrcpy(url, (size_t) (eof_protocol - url), 1);
 	/* log_debug(__func__, "Protocol: %s", (*result)->protocol); */
 
 	if (strlen((*result)->protocol) + 3 == strlen(url)) {
@@ -109,10 +109,10 @@ int hurl_parse_url(char *url_ptr, HURLParsedURL **result) {
 		log_debug(__func__, "Empty hostname.");
 		return 0;
 	}
-	(*result)->hostname = hurl_allocstrcpy(bgof_hostname, eof_hostname - bgof_hostname, 1);
+	(*result)->hostname = hurl_allocstrcpy(bgof_hostname, (size_t) (eof_hostname - bgof_hostname), 1);
 
 	if (bgof_path) {
-		(*result)->path = hurl_allocstrcpy(bgof_path, eof_url - bgof_path, 1);
+		(*result)->path = hurl_allocstrcpy(bgof_path, (size_t) (eof_url - bgof_path), 1);
 	} else {
 		(*result)->path = hurl_allocstrcpy("/", 1, 1);
 	}
